fix: Properly set BaseSHA on activities
First, make sure we set it even if we're not doing status reporting,
because that's just cleaner.

But more importantly, the preexisting logic would only set the
`BaseSHA` on the run info (and from there the activity) if the last
commit SHA wasn't already set by the presence of `PULL_PULL_REF` or
`PULL_BASE_REF` environment variables. Which...are actually present
now in Lighthouse, via
https://github.com/jenkins-x/lighthouse/issues/542. So let's handle
that correctly, and consistently with when we're on Prow.

fixes https://github.com/jenkins-x/lighthouse/issues/547

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>