IGNORE: Trying to verify node-http quickstart
It's failing on the boot-lh context on
https://github.com/jenkins-x/jenkins-x-versions/pull/866, which is
currently the only place we're calling it. I think I've got it working
on a testbed cluster I've been using, but I'm not confident enough to
be sure. I know I've got it failing on the github context on
https://github.com/jenkins-x/lighthouse/pull/589, but I want to verify
that it passes with Prow, so here's this PR.

The bdd and tekton contexts are tricked into not doing anything here,
and `boot-vault` is just running `test-single-import` with
`node-http`, so let's see what happens...

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>