fix: Add expiration time for token to UserAuth when using k8s secrets
This will allow for `GitProvider` implementations to check if the
token they're configured with has expired, and if so, reload their
auth config. I'll have a follow-up supporting this for GitHub
shortly. As mentioned on the issue, we need to eventually support
defining expiration time with Vault-stored git auth secrets as well,
but I wanted to start with the simplest proof of concept.

fixes #6963

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>