fix: Lower-case dir placeholder replacements from build pack
We lower-case the app name etc when generating directories (like
`charts/(app name)`) in `jx import`, but had not been doing so when
replacing `REPLACE_ME_APP_NAME` etc placeholders in directories
configured in build packs. We should be using the same case there that
we use in generating the directories in the first place.

fixes #6983

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>