chore: Disable PR status reporting if new enough Lighthouse present
This is preperatory for
https://github.com/jenkins-x/lighthouse/pull/603, which, among other
things, introduces a new deployment, `lighthouse-foghorn`, which will
handle SCM provider status reporting. While we'll update
`jenkins-x-boot-config` to not configure the build controller to do
status reporting when that Lighthouse PR merges, I'd like to also add
a check to see if the `lighthouse-foghorn` deployment is present and
skip reporting if it is. This will prevent potential issues with
anyone using master of the version stream picking up the new
Lighthouse but not the change to `jenkins-x-boot-config`.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>