chore: Bump antham/chyle to 1.6.0
This is part of #6484's prep work - antham/chyle 1.4.0 (and 1.5.0)
depend on an arbitrary commit in spf13/cobra, and that ends up
resulting in breaking dependency resolution in go 1.13. I initially
worked around this by adding an explicit `require` for spf13/cobra
0.0.5, but antham/chyle 1.6.0 passes all our tests and depends on
spf13/cobra 0.0.5 explicitly, so it's a better fix.

I also did a `go mod tidy` while I was here.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>