fix: Turn off Athens proxy when running "jx step create pr go"
This gets around the problem of `v2.0.1285` not resolving to a
pseudoversion with go 1.13, which the version of Athens we're now
running is using behind the scenes. The underlying problem is #7069,
and the long-term fix for that is to go to a proper semver usage, but
this should unbork things in the meantime.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>