fix: Handle Lighthouse-specific HMAC secret in update_webhooks
Sometimes, but weirdly not always, when switching from Prow to
Lighthouse, the `hmac-token` secret gets deleted. Currently, Prow and
Lighthouse use the same name for that secret, so I believe that
Lighthouse installs and then Prow gets removed, resulting in
Lighthouse not needing to create the secret, since it's identical to
the Prow one, but then the removal of Prow deleting the secret (and
possibly the `hook` service - still trying to verify that).

As I said, this doesn't happen consistently - I've now switched a
cluster from Prow to Lighthouse, back to Prow, and back to Lighthouse,
and the first time, Prow didn't get deleted at all, then the secret
stuck around after the installation of Prow and removal of Lighthouse,
and stuck around again after the installation of Lighthouse and
removal of Prow. So...I don't know if this is just random bad luck or what.

Regardless, I'm going to change Lighthouse to use a unique secret
name, but `jx update webhooks` needs to know that secret name. So
let's have it try the Prow name, and then if that's not found, try the
new Lighthouse name

part of fixing #7154

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>