feat: Add buildpack info to requirements
When we moved JX over to boot and Helm 2, a regression was introduced
where `jx edit buildpack` stopped working. Because JX team settings were
no longer used to populate the team settings for a CRD. So the default
buildpack info from the jx-bootresources values.yaml file were used.

This commit introduces the build pack config to the requirements file,
which will allow our users to specify custom build packs. We will
deprecate the `jx edit buildpack` command in a future commit.

Fixes #7126 and #6374