fix: Don't use pr:from_ref_updated webhook on older BitBucket Servers
`pr:from_ref_updated` was added in BitBucket Server 7.0.0. Without it,
we're never notified when a new commit is pushed to an existing PR, so
it's obviously something we really want enabled, but it doesn't exist
prior to 7.0.0. If we try to enable it on older versions, the whole
API call will fail. So let's only enable it if the target server is
new enough.

fixes #7204

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>