fix: making sure PVS finalizers get  set to nil during 'jx uninstall'
When uninstalling `jx`, the PVC would hang in a terminating state. When
we set the PVC, finalizer volume to `null`, the uninstall process
completes as normal.

There is an open issue in
[Kubernetes](https://github.com/kubernetes/kubernetes/issues/77258).

This commit sets the PVC finalizer to `null`, which fixes our issue and
allows users to uninstall `jx`.

Also adding `jx uninstall` as command to ci.sh of tf-boot context