fix: Read the entire reader before writing to bucket
For some reason, `io.Copy` was resulting in a "short write" error in
the build controller, and the best suspicion I've got at this point is
that it's a problem with the bucket writer closing early or something
along those lines. So let's instead go back to earlier behavior and
read the reader contents and write that directly rather than relying
on `io.Copy`.

fixes #7461 (hopefully!)

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>