fix: Increase max age of PipelineRuns for GCing
In practice, this results in builds that take longer than the max age
sometimes (though not always) failing to get their status actually
recorded properly. So let's increase that age, and also add a check to
make sure the `PipelineRun` is in a terminal state before we even
consider deleting it.

fixes #7444

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>