fix: step next-version and helm 3
fixes #7564

This makes the `jx step next-version` command compatible with Helm 3 when it reads/writes `Chart.yaml` files.

When the `jx step next-version` command reads the version from a Chart.yaml file, it looks for the `version` string, reads the value and stops there.
But when it writes a version in a Chart.yaml file, it replaces all the values following the `version` string.
This was not an issue with Helm 2, because there used to be a single `version` field in the Chart.yaml file.
But in Helm 3, the dependencies have been moved inside the Chart.yaml file, which means that each dependency now has a `version` field.
As a result, when the `jx step next-version` updates a version, it replaces all versions (including the dependencies) with the new value - which is incorrect.
This commit fixes this behaviour by stopping after the first replacement. It expect people to write the main chart version before the dependencies.

Signed-off-by: Vincent Behar <v.behar@free.fr>